vfs0090: retry scanning if failed because of read errors
Unfortunately it seems that fprintd (the only one acting like this
as libfprint clients, but the most important) requires the drivers
themselves to retry when they emit the retry-callback.

So, we do it in this driver too, although it currently breaks
some other tools in the examples or fprint_demo.

However, I don't really like this way of acting, as I'd expect
the caller to retry, and not the driver itself.